不能增加文件，改成内部类
