--- src/main/java/com/github/hcsp/io/GitHubPullResponse.java
@@ -1,43 +1,0 @@
-package com.github.hcsp.io;
-
-class GitHubPullResponse {
-    int number;
-    String title;
-    User user;
-
-    public int getNumber() {
-        return number;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setNumber(int number) {
-        this.number = number;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    static class User {
-        String login;
-
-        public String getLogin() {
-            return login;
-        }
-
-        public void setLogin(String login) {
-            this.login = login;
-        }
-    }
-}
